refactor: payment - presentation 레이어 패키지 정리#47
- mapper 위치 변경 (dto 하위)
- exception 위치 변경 (infrastructure 하위)
