+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/presentation/dto/PaymentPresentationMapper.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.presentation;
+package com.boeingmerryho.business.paymentservice.presentation.dto;
 
 import org.mapstruct.Mapper;
 import org.springframework.data.domain.Pageable;
